% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Compute your body mass index}
\usage{
bmi(sex, height, weight)
}
\description{
This function computes your body mass index(bmi) and shows your status through a graph.
}
\examples{

bmi('female', 166, 53)
}
